Adding a password system for notes

+++ src/components/OpenNote.tsx
@@ -11,6 +11,9 @@ const OpenNote: React.FC = () => {
   const [noteContent, setNoteContent] = useState<string>("");
   const [revealed, setRevealed] = useState<boolean>(false);
   const [noteNotFound, setNoteNotFound] = useState<boolean>(false);
+  const [password, setPassword] = useState<string>("");
+  const [storedPassword, setStoredPassword] = useState<string>("");
+  const [requiresPassword, setRequiresPassword] = useState<boolean>(false);
 
   useEffect(() => {
     if (noteId) {
@@ -24,7 +27,7 @@ const OpenNote: React.FC = () => {
     try {
       const { data, error } = await supabase
         .from("privnote")
-        .select("value")
+        .select("value, note_password")
         .eq("note_uid", noteId)
         .single();
 
@@ -32,6 +35,8 @@ const OpenNote: React.FC = () => {
         setNoteNotFound(true);
       } else {
         setNoteContent(data?.value ?? "");
+        setStoredPassword(data?.note_password ?? "");
+        setRequiresPassword(data?.note_password !== "");
       }
     } catch (error) {
       if (error instanceof Error) {
@@ -65,29 +70,59 @@ const OpenNote: React.FC = () => {
   }, [toasts]);
 
   const revealNote = async (): Promise<void> => {
-    try {
-      setRevealed(true);
-      if (noteId) {
-        await supabase.from("privnote").delete().eq("note_uid", noteId);
+    if (!requiresPassword || password === storedPassword) {
+      try {
+        setRevealed(true);
+        if (noteId) {
+          await supabase.from("privnote").delete().eq("note_uid", noteId);
+        }
+      } catch (error) {
+        if (error instanceof Error) {
+          console.error(error.message);
+        } else {
+          console.error("Error:", error);
+        }
+        toast.error("Failed to reveal note!");
       }
-    } catch (error) {
-      if (error instanceof Error) {
-        console.error(error.message);
-      } else {
-        console.error("Error:", error);
-      }
-      toast.error("Failed to reveal note!");
+    } else {
+      toast.error("Incorrect password!");
     }
   };
 
   const revealButton = (
     <button
       onClick={revealNote}
-      className="absolute inset-0 flex items-center justify-center w-full h-full overflow-hidden text-white transition-opacity bg-black bg-opacity-50 hover:bg-opacity-70"
+      className="relative flex items-center justify-center w-full h-full p-4 overflow-hidden text-white transition-opacity bg-black bg-opacity-50 hover:bg-opacity-70"
     >
-      <div className="absolute p-32 rounded-full hover:border-2 animate-ping"></div>
+      <div className="absolute p-32 rounded-full hover:border-2 animate-ping "></div>
       <span className="text-xl font-semibold">Reveal Note</span>
     </button>
+  );
+
+  const passwordInput = (
+    <div className="flex flex-col items-center w-full">
+      <button
+        onClick={revealNote}
+        className="relative flex items-center justify-center w-full h-full p-4 py-10 mb-5 overflow-hidden text-white transition-opacity bg-black bg-opacity-50 hover:bg-opacity-70"
+      >
+        <div className="absolute p-32 rounded-full hover:border-2 animate-ping "></div>
+        <span className="text-xl font-semibold">Reveal Note</span>
+      </button>
+   
+        <p>Enter password to unlock note</p>
+        <input
+          type="password"
+          placeholder="type here..."
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+          className={`w-40 px-3 py-2 my-2 border rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 ${
+            darkMode
+              ? "text-white bg-stone-800 border-gray-700"
+              : "text-gray-700 bg-white border-gray-300"
+          }`}
+        />
+    
+    </div>
   );
 
   const textArea = (
@@ -141,10 +176,9 @@ const OpenNote: React.FC = () => {
         )
       ) : (
         <>
-          <div className="relative w-full p-4 sm:w-4/5 md:w-3/5 xl:w-2/5 h-1/4">
-            <div className="absolute inset-0 bg-blur" />
-            {revealButton}
-            <p className="text-center text-gray-300 bottom-2 left-2 animate-pulse">
+          <div className="flex flex-col items-center w-full p-4 sm:w-4/5 md:w-3/5 xl:w-2/5 h-1/4">
+            {requiresPassword ? passwordInput : revealButton}
+            <p className="mt-4 text-center text-gray-300 animate-pulse">
               You can only view this note once.
             </p>
           </div>
